Added support for the XZR, Z1C and XT.
XP and XJ added but not functional yet, a CM repo issue has to be solved first.

Updated dualrecovery.sh to work with a reasonably big change: Before this version it would try to clean up the kernel ramdisk before unpacking the recovery ramdisk. Instead it will now create a folder, unpack everything there and then chroot in to that folder executing the recovery's init. This prevents accidental wiping of partitions or mount locations XZDualRecovery does not know about and does not unmount before it starts recovery.
By default, all previous versions would copy all the recovery files from /system/bin to /sbin. It will now only copy the one which it will be using. This speeds up booting a bit, which is always good.

Updated the rickiller script to hopefully fix the remount issue on the Xperia ZR.

PhilZ will now have a correctly sized button bar on the Xperia T, Xperia ZR and Xperia Z1 Compact.

Hopefully fixed the TWRP theme on the Xperia ZR.