Now with installation option to install on an unrooted device, works only on specific FW versions. The feature uses rootkitXperia for this purpose, adding it to the repo with this commit.
Changed the event input node search engine. It now searches once and adds the nodes to the XZDR.prop file and dualrecovery.sh will now read dr.gpiokeys.node and dr.pwrkey.node for them to use in the key checking process.

Updated the search engine for the power button, the old one was not able to find the power key on the Xperia Z1.