Updated PhilZ to the latest build.
Now that there are official builds for 720x1280 devices, i've started using the Xperia SP version to repack for the other non-officially supported devices.

Change to skip the keycheck on the boot to recovery file and when the reboot recovery trigger is found, proposed by Shoey63.

Attempted to fix the recovery.fstab for the Xperia SP and T/V/TL/TX (not entirely working yet)