redone xzdualrecovery.sh, it now assumes the recovery.fstab will be correctly configured and it will be able to mount system and userdata. For some reason it still causes a reboot on the Xperia T family... I'm not sure why.
redone my CM repo folder structure... adapted the build suite accordingly.
This has been done because I kept running in to problems while trying to build any other branch on the CyanogenMod repo's...
As a result I hope to be able to build a version for the Xperia P and J soon.

Redesigned DRGETPROP and DRSETPROP to work more precise, it was not able to differentiate between an unset but present variable or a variable which was completely missing.
DRGETPROP will now return 'false' when a variable/property set is completely missing from the properties file or an empty value for when it's present but unset, as a result
DRSETPROP will set the value on the existing line in the properties file, regardless of it's current value or add a new one. This now works ok, all the time.

Installation of XZDR now sets the current version and release stream on the properties file, for furure use and reference to what version has been installed.

Tried to fix recovery.fstab on the Xperia T family again... unsuccessfully ... this also spawned the need for a seperated repo structure.

The installation routine now also creates /system/.XZDualRecovery and copies busybox in to it. Once a user installs a Play busybox and breaks compatibility because of missing
lzma functionality chargemon.sh will now nolonger crap out and stop operation but copies the required busybox back to /system/xbin.
This change was getting more important by the second as the Z1, Z1 Compact and Z Ultra with their JB4.3 built-in kernel protection of system and rootfs will lose full root once
that happens, making it impossible to restore a good busybox...