Start work on the XZDRKernel for the Xperia T, needs a different tool to create the image.
During installation and while flashing an update it will now search for the event nodes to update them.

This change is required to make updating from one ROM (with event9 as gpio key event node) to another (with event6 as node) possible, this way the users will not have to remove the XZDR.prop file to regenerate it.

setversion.sh should now be renamed to be honest...