xzdualrecovery.sh: Changed the way system and data are mounted to get it working correctly on the Xperia T. patches/xt/twrp/etc/twrp.fstab: removed the extra internal storage mention...
the rest of the ramdisks patched with the latest /tmp fix.

dualrecovery.sh: NOT unmounting sys and proc to keep the Xperia T's modem from throwing a fit and causing a reboot while in recovery.

installrecovery.sh/setversion.sh: the keys node update code debugged, it was storing a false name in the properties file.