Modified postrecoveryboot.sh to not throw an error anymore if there is no reason for it to throw it.
Updated DRSET-/GET-PROP to return null instead of false. If a switch was set to false it would add it to the config as if it wasn't there...

Updated TWRP to 2.8.6.0 again, because all other issues have been solved, this might have a positive effect on the TWRP operation too.