Removed chargemon executables, these where still included but are really, really old...
Changed the code for the hijack backup of the stock binary. It will check if the stock file is there and if the chargemon file in system/bin is a binary or a script before moving it away to the backup file.

Removed some smaller redundant stuff, also tried to improve the failed installation detection.