New kernel for Z2 (5.1.1 kernel).
Reverted the modulecrcpatch binary to the old/original version on the kernel ramdisk patches to fix issues caused BY the PIE nature of this file... who would have thought!

The kernel version now stores the byeselinux module on multiple locations, to have a failsafe. This way the module will be available even if the system partition is inaccessible.